# docs/swagger.yaml
definitions:
  subscriptions.Subscription:
    description: Модель подписки
    properties:
      end_date:
        example: 01-2001
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        example: 01-2001
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
  description: API для агрегации записей о подписках
  title: subscriptions API
  version: "1.0"
paths:
  /api/subscriptions:
    get:
      consumes:
      - application/json
      description: Возвращает все имеющиеся записи о подписках
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/subscriptions.Subscription'
            type: array
        "400":
          description: '''error'': ''message'''
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '''error'': ''message'''
          schema:
            additionalProperties: true
            type: object
      summary: Получить все записи о подписках
      tags:
      - Subscriptions
    post:
      consumes:
      - application/json
      description: Создает новую запись о подписке
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/subscriptions.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Запись о подписке успешно создана
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '''error'': ''message'''
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '''error'': ''message'''
          schema:
            additionalProperties: true
            type: object
      summary: Создать запись о подписке
      tags:
      - Subscriptions
  /api/subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет запись о подписке по id
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Задача успешно удалена
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '''error'': ''message'''
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '''error'': ''message'''
          schema:
            additionalProperties: true
            type: object
      summary: Удалить запись о подписке
      tags:
      - Subscriptions
    get:
      consumes:
      - application/json
      description: Возвращает данные о подписке по ее id
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions.Subscription'
        "400":
          description: '''error'': ''message'''
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '''error'': ''message'''
          schema:
            additionalProperties: true
            type: object
      summary: Получить данные о подписке
      tags:
      - Subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет данные о подписке по ее id
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/subscriptions.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: Запись о подписке успешно обновлена
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '''error'': ''message'''
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '''error'': ''message'''
          schema:
            additionalProperties: true
            type: object
      summary: Обновить данные о подписке
      tags:
      - Subscriptions
swagger: "2.0"